Move BlockTemplate and AssemblerOptions to their own files
